+++ src/InternalForces/Actuators/ActuatorGauss3p.cpp
@@ -57,7 +57,7 @@ internal_forces::actuator::ActuatorGauss3p::ActuatorGauss3p(
     const utils::Scalar& w1,
     const utils::Scalar& r,
     const utils::Scalar& qopt,
-    unsigned int dofIdx) :
+    size_t dofIdx) :
     internal_forces::actuator::Actuator(direction, dofIdx),
     m_k(std::make_shared<utils::Scalar>(4.3)),
     m_Tmax(std::make_shared<utils::Scalar>(Tmax)),
@@ -85,7 +85,7 @@ internal_forces::actuator::ActuatorGauss3p::ActuatorGauss3p(
     const utils::Scalar& w1,
     const utils::Scalar& r,
     const utils::Scalar& qopt,
-    unsigned int dofIdx,
+    size_t dofIdx,
     const utils::String &jointName) :
     internal_forces::actuator::Actuator(direction, dofIdx, jointName),
     m_k(std::make_shared<utils::Scalar>(4.3)),
@@ -143,8 +143,8 @@ utils::Scalar internal_forces::actuator::ActuatorGauss3p::torqueMax(
     const rigidbody::GeneralizedCoordinates &Q,
     const rigidbody::GeneralizedVelocity &Qdot)
 {
-    utils::Scalar pos(Q[*m_dofIdx] * 180/M_PI);
-    utils::Scalar speed(Qdot[*m_dofIdx] * 180/M_PI);
+    utils::Scalar pos(Q[static_cast<unsigned int>(*m_dofIdx)] * 180/M_PI);
+    utils::Scalar speed(Qdot[static_cast<unsigned int>(*m_dofIdx)] * 180/M_PI);
 
     // Tetanic torque max
     utils::Scalar Tc = *m_T0 * *m_wc / *m_wmax;
